[Step5] refactor : 코드리뷰 - 랜덤 로직을 포함한 테스트 케이스 추가
