+++ src/test/kotlin/racing/domain/car/CarTest.kt
@@ -1,10 +1,9 @@
-package racing
+package racing.domain.car
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
-import racing.domain.car.Car
 import racing.domain.exception.CarNameLengthExcessException
 
 class CarTest {
