+++ src/test/kotlin/racing/domain/race/WinnerJudgeTest.kt
@@ -1,9 +1,8 @@
-package racing
+package racing.domain.race
 
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import racing.domain.common.CarRaceProperty
-import racing.domain.race.WinnerJudge
 
 class WinnerJudgeTest {
